+++ src/Data/Demand.php
@@ -7,19 +7,22 @@ use Becklyn\SchemaOrg\TypeChecker\TypeChecker;
 class Demand extends Intangible
 {
     //region Fields
-    private ?string $areaServed = null;
+    /** @var string|GeoShape */
+    private $areaServed;
     private ?string $availability = null;
     private ?\DateTimeImmutable $availabilityEnds = null;
     private ?\DateTimeImmutable $availabilityStarts = null;
     private ?Place $availableAtOrFrom = null;
     private ?string $businessFunction = null;
-    private ?string $eligibleRegion = null;
+    /** @var string|GeoShape */
+    private $eligibleRegion;
     private ?string $gtin = null;
     private ?string $gtin12 = null;
     private ?string $gtin13 = null;
     private ?string $gtin14 = null;
     private ?string $gtin8 = null;
-    private ?string $ineligibleRegion = null;
+    /** @var string|GeoShape */
+    private $ineligibleRegion;
     private ?QuantitativeValue $inventoryLevel = null;
     private ?OfferItemCondition $itemCondition = null;
     private ?PriceSpecification $priceSpecification = null;
@@ -35,7 +38,10 @@ class Demand extends Intangible
 
 
     //region Accessors
-    public function getAreaServed () : ?string
+    /**
+     * @return GeoShape|string
+     */
+    public function getAreaServed ()
     {
         return $this->areaServed;
     }
@@ -71,7 +77,10 @@ class Demand extends Intangible
     }
 
 
-    public function getEligibleRegion () : ?string
+    /**
+     * @return GeoShape|string
+     */
+    public function getEligibleRegion ()
     {
         return $this->eligibleRegion;
     }
@@ -107,7 +116,10 @@ class Demand extends Intangible
     }
 
 
-    public function getIneligibleRegion () : ?string
+    /**
+     * @return GeoShape|string
+     */
+    public function getIneligibleRegion ()
     {
         return $this->ineligibleRegion;
     }
@@ -165,13 +177,19 @@ class Demand extends Intangible
     //endregion
 
 
+    //region Withers
     /**
+     * @param string|GeoShape $areaServed
+     *
      * @return static
      */
-    public function withAreaServed (?string $areaServed)
+    public function withAreaServed ($areaServed)
     {
+        TypeChecker::ensureIsValidValue($areaServed, TypeChecker::OPTIONAL, "string", GeoShape::class);
+
         $clone = clone $this;
         $clone->areaServed = $areaServed;
+
         return $clone;
     }
 
@@ -232,12 +250,17 @@ class Demand extends Intangible
 
 
     /**
+     * @param string|GeoShape $eligibleRegion
+     *
      * @return static
      */
-    public function withEligibleRegion (?string $eligibleRegion)
+    public function withEligibleRegion ($eligibleRegion)
     {
+        TypeChecker::ensureIsValidValue($eligibleRegion, TypeChecker::OPTIONAL, "string", GeoShape::class);
+
         $clone = clone $this;
         $clone->eligibleRegion = $eligibleRegion;
+
         return $clone;
     }
 
@@ -298,12 +321,17 @@ class Demand extends Intangible
 
 
     /**
+     * @param string|GeoShape|null $ineligibleRegion
+     *
      * @return static
      */
-    public function withIneligibleRegion (?string $ineligibleRegion)
+    public function withIneligibleRegion ($ineligibleRegion)
     {
+        TypeChecker::ensureIsValidValue($ineligibleRegion, TypeChecker::OPTIONAL, "string", GeoShape::class);
+
         $clone = clone $this;
         $clone->ineligibleRegion = $ineligibleRegion;
+
         return $clone;
     }